BUG: #1564 Integer overflows in linkage()
In cases where `linkage` is given large data sets, it has to request
`NCHOOSE2(n)` `double`s in memory. However, as `n` is an `int`, it will
overflow if it is large. The changes here simply declare three relevant
variables as `long` instead: `n`, `np`, and `npc2`. The latter two are simply
declared in the body of the function. The first one is an argument to the
function. However, the `linkage_wrap` function gives it an `int` at that
position, so it will have to perform an implicit cast.